ALso cast lat,lon to double here when adding to ptree

+++ src/loki/service.cc
@@ -165,8 +165,8 @@ namespace valhalla {
           shape_count = input_shape->size();
           for (auto& latlng : *input_shape) {
             auto ll = baldr::Location::FromPtree(latlng.second).latlng_;
-            latlng.second.put("lon", ll.first);
-            latlng.second.put("lat", ll.second);
+            latlng.second.put("lon", static_cast<double>(ll.first));
+            latlng.second.put("lat", static_cast<double>(ll.second));
           }
         }//compressed shape
         //if we receive as encoded then we need to add as shape to request
